Add example systemd configuration to handle the srcpd with systemctl
For systemd, the srcpd is not started in the implemented daemon mode.
This mode is obsolete anyway and prohibits starting and stopping the daemn by systemctl.

The daemon is started as root and needs root rights to setup serial port speed for DDL (=Digital Direct Loco).
But you can put another user/group in the `srcpd.conf` file, the daemon will switch to this user afterwards.